+++ pilot/pkg/xds/ads.go
@@ -16,7 +16,6 @@ package xds
 
 import (
 	"fmt"
-	"io"
 	"strconv"
 	"strings"
 	"sync/atomic"
@@ -130,23 +129,6 @@ func newConnection(peerAddr string, stream DiscoveryStream) *Connection {
 	}
 }
 
-// isExpectedGRPCError checks a gRPC error code and determines whether it is an expected error when
-// things are operating normally. This is basically capturing when the client disconnects.
-func isExpectedGRPCError(err error) bool {
-	if err == io.EOF {
-		return true
-	}
-
-	s := status.Convert(err)
-	if s.Code() == codes.Canceled || s.Code() == codes.DeadlineExceeded {
-		return true
-	}
-	if s.Code() == codes.Unavailable && s.Message() == "client disconnected" {
-		return true
-	}
-	return false
-}
-
 func (s *DiscoveryServer) receive(con *Connection, reqChannel chan *discovery.DiscoveryRequest, errP *error) {
 	defer func() {
 		close(reqChannel)
@@ -162,7 +144,7 @@ func (s *DiscoveryServer) receive(con *Connection, reqChannel chan *discovery.Di
 	for {
 		req, err := con.stream.Recv()
 		if err != nil {
-			if isExpectedGRPCError(err) {
+			if istiogrpc.IsExpectedGRPCError(err) {
 				log.Infof("ADS: %q %s terminated %v", con.PeerAddr, con.ConID, err)
 				return
 			}
